+++ js/main.js
@@ -1,2 +1,2 @@
 import './popup.js';
-import './form.js';
+import './form-validation.js';
